server/pyproject.toml
[project]
name = "bytebuddhi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "hiredis>=3.3.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langsmith>=0.6.4",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqlalchemy[mypy]>=2.0.46",
    "sse-starlette>=3.2.0",
    "uvicorn[standard]>=0.40.0",
    "python-dotenv>=1.2.1",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
    "httpx>=0.28.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "gitpython>=3.1.46",
    "tiktoken>=0.12.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "anthropic>=0.76.0",
    "python-dateutil>=2.9.0.post0",
    "orjson>=3.11.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "black>=26.1.0",
    "ruff>=0.14.14",
    "mypy>=1.19.1",
    "isort>=7.0.0",
    "types-redis>=4.6.0.20241004",
    "types-passlib>=1.7.7.20250602",
    "ipython>=9.9.0",
    "ipdb>=0.13.13",
    "watchdog>=6.0.0",
    "greenlet>=3.3.1",
]
